+++ src/business-logic/OptionsLogic.ts
@@ -1083,7 +1083,8 @@ export class OptionsLogic {
       this.isFullscreenMode,
     );
     const textColor = getCurrentValue(
-      this.options?.styling?.toolbar?.progressBar?.screenshotViewer?.textOrForegroundColor,
+      this.options?.styling?.toolbar?.progressBar?.screenshotViewer
+        ?.textOrForegroundColor,
       this.theme.colorFive,
       this.isFullscreenMode,
     );