+++ src/business-logic/StylingLogic.ts
@@ -670,9 +670,11 @@ export class StylingLogic {
 
     const widthStyle = {
       width: widthToUse,
-      maxWidth: this.isFullscreenMode ? undefined : `calc(100% - ${
-        CAROUSEL_ITEM_SPACING_DEFAULT * 2
-      }${CAROUSEL_SPACING_UNIT})`,
+      maxWidth: this.isFullscreenMode
+        ? undefined
+        : `calc(100% - ${
+            CAROUSEL_ITEM_SPACING_DEFAULT * 2
+          }${CAROUSEL_SPACING_UNIT})`,
       boxShadow: `0 10px 15px -3px rgba(0,0,0,.25)`,
     } as CSSProperties;
     const paddingStyle = isMinimized
@@ -1158,7 +1160,8 @@ export class StylingLogic {
   }
 
   get carouselVideoProgressScreenshotViewerTextContainerStyle() {
-    const { width, textColor } = this.optionsLogic.videoProgressBarScreenshotViewer;
+    const { width, textColor } =
+      this.optionsLogic.videoProgressBarScreenshotViewer;
     return {
       color: textColor,
       position: "absolute",